Consolidate all the parsers into `parsers`
